# resources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Norevify API",
    "version": "1.0.0",
    "description": "Vehicle expense tracking API with JWT authentication. All authenticated endpoints require `Authorization: Bearer <token>`. Some endpoints additionally require a verified email (`withVerifiedUser`) or the `pro` plan."
  },
  "servers": [{ "url": "/" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static secret from the CRON_SECRET environment variable, not a user JWT."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Error description" }
        }
      },
...
      "ProfitSummary": {
        "type": "object",
        "properties": {
          "period": {
            "type": "object",
            "properties": {
              "year": { "type": "integer", "example": 2026 },
              "month": { "type": "integer", "example": 4 }
            }
          },
          "totalIncome": { "type": "number", "example": 2400.0 },
          "totalExpenses": { "type": "number", "example": 900.0 },
          "profit": { "type": "number", "example": 1500.0 },
          "profitPerKm": { "type": "number", "nullable": true, "example": 1.2 },
          "profitPerDay": { "type": "number", "example": 50.0 },
          "kmDriven": {
            "type": "number",
            "nullable": true,
            "description": "Fill-to-fill odometer span used for the fuel cost calculation",
            "example": 1250
          },
          "hours": { "type": "number", "nullable": true, "example": 2.95 },
          "workKm": { "type": "number", "nullable": true, "example": 41.0 },
          "deliveries": { "type": "integer", "nullable": true, "example": 7 },
          "costPerKm": {
            "type": "number",
            "nullable": true,
            "description": "Fill-to-fill fuel DKK/km; null until 2+ odometer-tagged fill-ups exist",
            "example": 1.35
          },
          "costPerKmSamples": {
            "type": "integer",
            "example": 2,
            "description": "Number of fill-ups used in costPerKm (excludes the anchor fill)"
          },
          "fuelCost": { "type": "number", "nullable": true, "example": 55.35 },
          "netEarnings": { "type": "number", "nullable": true, "example": 339.02 },
          "netPerHour": { "type": "number", "nullable": true, "example": 114.92 },
          "netPerKm": { "type": "number", "nullable": true, "example": 8.27 },
          "grossPerHour": { "type": "number", "nullable": true, "example": 133.68 },
          "months": {
            "type": "array",
            "description": "Present only when `breakdown=monthly` was requested and no `month` filter was given. One profit row per calendar month of `period.year`.",
            "items": {
              "type": "object",
              "properties": {
                "month": { "type": "integer", "example": 4 },
                "totalIncome": { "type": "number", "example": 500.0 },
                "costPerKm": { "type": "number", "nullable": true, "example": 1.35 },
                "costPerKmSamples": { "type": "integer", "example": 2 },
                "fuelCost": { "type": "number", "nullable": true, "example": 55.35 },
                "netEarnings": { "type": "number", "nullable": true, "example": 339.02 },
                "netPerHour": { "type": "number", "nullable": true, "example": 114.92 },
                "hours": { "type": "number", "nullable": true, "example": 2.95 },
                "workKm": { "type": "number", "nullable": true, "example": 41.0 }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token not provided or invalid/expired",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  },
  "paths": {
...
    "/api/income/summary": {
      "get": {
        "tags": ["Income"],
        "summary": "Get income vs. expenses profit summary for a period",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "example": 2026
          },
          {
            "name": "month",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 12 },
            "example": 4
          },
          { "name": "vehicleId", "in": "query", "schema": { "type": "string" } },
          {
            "name": "breakdown",
            "in": "query",
            "description": "Pass 'monthly' (with no month param) to also return a `months` array — one profit row per calendar month of `year`.",
            "schema": { "type": "string", "enum": ["monthly"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Profit summary",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfitSummary" } }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}